feat: add lsusb crate and list devices example
Add the list devices example. I still need to think of a good way of
getting all the devices that are "USB keyboards".

I already have a check to do so, using the "sys/bus" interfaces, but it
isn't readable and I don't know whether it works on different OSs.