refactor: add find_for_serial_ids and find_by_serial_id
Both functions are expected to do the same. I expect the only
interesting one to be find_by_serial_id, as maybe we should enforce only
using one keyboard. Using multiple keyboards may be interesting, but how
do we take the fact that a user could use different keyboards connected
to their machine? This could be a segway into thinking about knowing the
type of the keyboard with a keypress. The following questions arise:

- Should layout_switcher change the layout if using multiple keyboards
connected at the same time?
- Should the scope only be to switch layouts on usb connect/disconnect?