test: add start of tests for find_by_serial_id funcs
Testing structs is not easy, considering we pass the static type to the
function that we have implemented. It is key to test the output of our
functions with a mock `Enumerator`, which has a defined set of
devices. Each person running the binary could have different devices and
therefore make local tests fail.