feat: add some notes, remove dependencies
Not looking at testing, still learning the language and uning mocking
libraries is too hard atm. Will come back for sure.

Add notes in the readme, I find the thought process to be
interesting.

Add some code to test the Command struct, and convert the stdout result
to a String (or &str).