feat: change &str for String and add json config
We use the config.json file under the .config/layout_switcher directory
for our deserialization purposes. We can use the file or future
arguments to override the config in the file.

I have been thinking about the priority of the arguments, and I believe
it should be:
1. If the keyboard in the arguments is found, modify the
LayoutSwitcherConfig.keyboards HashMap.
2. Else, add the keyboard on top priority, as it will be the keyboard to
be used.