feat: add cli args and its logic
What we do is that we take the value that is passed in the command line
args and if present in the config file we update it. If not present, it
should be the keyboard to be used, so should be put at the beginning of
the hashmap. I'm still not sure how to do so.