feat: add the benchmarking code
I have been working on improving the resources needed to run this in the
background. Using the libusb lib (which also has the advantage of
running on macos and windows), we have a really low exec time and therefore
low resources. Here are the results,

find_with_libusb: 26.67µs
find_with_udev: 135.13ms

Using the libusb library, even the dev process running uses:
%CPU %MEM CMD
 0.7  0.0 target/debug/layout_switcher monitor

Compared to the previous 20 or 30 percent, this is a great improvement.